+++ src/vscode.proposed.customEditorDiffs.d.ts
@@ -1,6 +1,8 @@
 /*---------------------------------------------------------------------------------------------
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *  Source: https://github.com/microsoft/vscode/blob/1.133.0/src/vscode-dts/vscode.proposed.customEditorDiffs.d.ts
+ *  Target: VS Code 1.133.0
  *--------------------------------------------------------------------------------------------*/
 
 declare module 'vscode' {